+++ el_pagination/loaders.py
@@ -1,19 +1,14 @@
 """Django EL Pagination object loaders."""
 
-from __future__ import unicode_literals
-
 from django.core.exceptions import ImproperlyConfigured
 
-try:
-    from importlib import import_module
-except ImportError:
-    from django.utils.importlib import import_module
+from importlib import import_module
 
 
 def load_object(path):
     """Return the Python object represented by dotted *path*."""
     i = path.rfind('.')
-    module_name, object_name = path[:i], path[i + 1:]
+    module_name, object_name = path[:i], path[i + 1 :]
     # Load module.
     try:
         module = import_module(module_name)