+++ el_pagination/tests/integration/test_callbacks.py
@@ -1,6 +1,6 @@
 """Javascript callbacks integration tests."""
 
-from __future__ import unicode_literals
+
 
 from el_pagination.tests.integration import SeleniumTestCase