+++ el_pagination/tests/integration/test_twitter.py
@@ -1,6 +1,6 @@
 """Twitter-style pagination integration tests."""
 
-from __future__ import unicode_literals
+
 
 from el_pagination.tests.integration import SeleniumTestCase