+++ el_pagination/tests/test_models.py
@@ -1,6 +1,6 @@
 """Model tests."""
 
-from __future__ import unicode_literals
+
 
 from contextlib import contextmanager