+++ el_pagination/tests/test_utils.py
@@ -1,6 +1,6 @@
 """Utilities tests."""
 
-from __future__ import unicode_literals
+
 
 from django.test import TestCase
 from django.test.client import RequestFactory