+++ el_pagination/tests/test_views.py
@@ -1,6 +1,6 @@
 """View tests."""
 
-from __future__ import unicode_literals
+
 
 from django.core.exceptions import ImproperlyConfigured
 from django.http import Http404