+++ tests/project/views.py
@@ -1,6 +1,6 @@
 """Test project views."""
 
-from __future__ import unicode_literals
+
 
 from django.shortcuts import render
 from django.views.generic import ListView